never create game as a matlab structure in the first place

+++ direction_get_angle.m
@@ -1,2 +1,2 @@
 function [angle] = direction_get_angle(direction)
-angle = direction*pi/2;
+angle = double(direction)*pi/2;
